Refactor app.py into modular flask layout
