+++ backend/pyproject.toml
@@ -17,6 +17,9 @@ package-mode = false
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.5"
 
+[tool.pytest.ini_options]
+pythonpath = "."
+
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
